Mega_Master Up to Spec
Brought the code up to function! It compiles with the Mega compiler, and is written to use all the hardware serials we're given. Since we're controlling 2 motors and 3 BLE chips over serial, our runtiest BLE sensor is left with a bonus software serial.

The code doesn't have the support for 3 BLE chips' worth of telemetry. Check out the internal 'update log' box for details.